qtbase: keep Qtbase synchronized with thud branch
For linux4sam 6.1 release, thud revision is used, but QT 5.12 is needed
for the Qt applications in meta-atmel. Since branch thud of meta-qt5
uses QT 5.11, we had to use 5.12 branch of meta-qt5 instead.

So, we need to apply a specific revision of '0004-Provide-access-to-linuxfb-dri-fd-through-platform.patch' file
for QT versions under 5.11.

Fixes #151

Signed-off-by: Pierre-Jean Texier <pjtexier@koncepto.io>